Separated Api and Web Controller for better maintainability
