+++ src/main/java/io/dankoller/codesharingplatform/controller/RequestController.java
@@ -5,21 +5,18 @@ import io.dankoller.codesharingplatform.persistence.CodeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDateTime;
 import java.util.*;
 
+// Api Controller
 @RestController
 public class RequestController {
 
     @Autowired
     private CodeService codeService;
-
-    /* API specific code below */
 
     // Return posted snippets via API
     @GetMapping("/api/code/{uuid}")
@@ -64,49 +61,4 @@ public class RequestController {
     public List<Code> getLatest() {
         return codeService.findTenRecentCodes();
     }
-
-    // Web Interface for the app
-    @Controller
-    class WebInterface {
-        // Main page
-        @GetMapping("/")
-        public String getIndexHtml(Model model) {
-            List<Code> list = codeService.findTenRecentCodes();
-            model.addAttribute("list", list);
-            return "index";
-        }
-
-        // Get the n-th uploaded code snippet
-        @GetMapping("/code/{uuid}")
-        public String getCodeHtml(Model model, @PathVariable String uuid) {
-
-            Code _code = codeService.findCodeByUuid(uuid);
-
-            if (_code == null) {
-                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
-            }
-
-            if (_code.isRestrictedByTime() || _code.isRestrictedByViews()) {
-                codeService.updateTimeAndViews(_code);
-            }
-
-            model.addAttribute("code", _code);
-            return "code";
-        }
-
-        // Return the 10 most recent snippet via HTML
-        @GetMapping("/code/latest")
-        public String getLatestHtml(Model model) {
-            List<Code> latest = codeService.findTenRecentCodes();
-            model.addAttribute("latest", latest);
-            return "latest";
-        }
-
-        // Post new code snippets page
-        @GetMapping("code/new")
-        public String getNewHtml() {
-            return "new";
-        }
-    }
-
 }
